+++ ImagesList/ImagesListCell.swift
@@ -1,0 +1,6 @@
+import UIKit
+
+final class ImagesListCell: UITableViewCell {
+    static let reuseIdentifier = "ImagesListCell"
+    
+}
